requirements.txt
# ===== LangChain Core =====
langchain
langchain-core
langchain-community
langchain-text-splitters

# ===== Web App =====
flask

# ===== OpenAI Integration =====
langchain-openai
openai
tiktoken

# ===== Anthropic Integration =====
langchain-anthropic

# ===== Google Gemini Integration =====
langchain-google-genai
google-generativeai

# ===== Hugging Face Integration =====
langchain-huggingface
transformers
huggingface-hub

# ===== Vector DB / RAG =====
faiss-cpu
youtube-transcript-api

# ===== Environment Variables =====
python-dotenv

# ===== Machine Learning Utilities =====
numpy
scikit-learn
